b2b_outreach_tool/requirements.txt
requests
beautifulsoup4
lxml
pyyaml
aiohttp
aiodns
flask
streamlit
pandas
python-dotenv
playwright
openai-whisper
torch
ffmpeg-python
setuptools-rust
psutil
